refactor: Edit splash screen module
