+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/presentation/AuthScreen.kt
@@ -134,7 +134,6 @@ internal fun AuthScreen(
             )
         }
     }
-    SplashScreen()
     if (showWebView) {
         StartKakaoLoginWebView(
             clientId = BuildConfig.KAKO_CLIENT_ID,
