+++ feature/main/src/main/java/com/teamoffroad/feature/main/MainActivity.kt
@@ -6,11 +6,15 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.annotation.RequiresApi
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.teamoffroad.core.designsystem.theme.OffroadTheme
 import com.teamoffroad.feature.main.component.MainTransparentActionBar
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.delay
 
 @AndroidEntryPoint
 @RequiresApi(Build.VERSION_CODES.TIRAMISU)
@@ -20,13 +24,22 @@ class MainActivity : ComponentActivity() {
 
         setContent {
             val navigator: MainNavigator = rememberMainNavigator()
+            val showSplash = remember { mutableStateOf(true) }
 
+            LaunchedEffect(Unit) {
+                delay(1100)
+                showSplash.value = false
+            }
             MainTransparentActionBar(window)
             OffroadTheme {
-                MainScreen(
-                    navigator = navigator,
-                    modifier = Modifier
-                )
+                if (showSplash.value) {
+                    SplashScreen()
+                } else {
+                    MainScreen(
+                        navigator = navigator,
+                        modifier = Modifier
+                    )
+                }
             }
         }
     }
